Sort facets by value
The "postcode" facetlist can have way more than 1400 items.
Returning the results at random leaves gaps between postcodes,
making it look like records are missing. This is confusing for endusers.
By sorting alphabetically the frontend can show a list without gaps,
with a warning on the bottom that not all facets are shown.